refactor: Rename sideCardData file
